% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmodelApply.R
\name{pmodelApply}
\alias{pmodelApply}
\title{Predict using a pmodel}
\usage{
pmodelApply(mymodel, exprData)
}
\arguments{
\item{mymodel}{An existing pmodel}

\item{exprData}{Expression data matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A vector of predicted efficacies.
}\if{html}{\out{</div>}}
}
\description{
pmodelApply takes an existing pmodel and applies it to an input set of samples.
}
